Enable the logs on Linux and make them run without error
Enable the logs and the memory leak reporter. The logs are now working, they can be found in a subdirectory of executable: ./logs. The memory leak reporter does not work, but it is already creating the leak report files, which can be found in the executable directory: CTP_LEAKS_99999.TXT, CTP_LEAKS_ALT_99999.TXT. So far these files do not contain any usable information, yet.
../ctp2_code/ctp/civ3_main.cpp

The logger tried to delete the directories logs/.. and logs/., which of course it should not try.
../ctp2_code/ctp/ctp2_utils/c3debug.cpp

Fixed a crash that occurred when creating the logs. And increased the buffer size for the debug assert messages, to avoid a stack overflow by writing in unallocated memory.
../ctp2_code/ctp/debugtools/log.cpp

+++ ctp2_code/ctp/ctp2_utils/c3debug.cpp
@@ -116,9 +116,13 @@ void c3debug_InitDebugLog()
 	while((dent = readdir(dir)))
 	{
 		int unlinkRetVal;
-		sprintf(fileName, "logs%s%s", FILE_SEP, dent->d_name);
-		unlinkRetVal = unlink(fileName);
-		Assert(unlinkRetVal == 0);
+		if(strcmp(dent->d_name, "." ) != 0
+		&& strcmp(dent->d_name, "..") != 0)
+		{
+			sprintf(fileName, "logs%s%s", FILE_SEP, dent->d_name);
+			unlinkRetVal = unlink(fileName);
+			Assert(unlinkRetVal == 0);
+		}
 	}
 
 	closedir(dir);
@@ -309,7 +313,7 @@ void c3debug_Assert(char const *s, char const * file, int line)
 	}
 	while (0);
 #else
-	MBCHAR str[64];
+	MBCHAR str[1024];
 	sprintf(str, "Assertion (%s) Failed in File:%s, Line:%ld\n", s, file, line);
 	fprintf(stderr, str);
 	sint32 result = MessageBox(NULL, str, "Assert", MB_ABORTRETRYIGNORE | MB_ICONEXCLAMATION);
